# Use OrthoDB SPARQL API for more robustness and speed