--- a/src/homology.js
+++ b/src/homology.js
@@ -1,11 +1,13 @@
 import fetchOrthologsFromOma from './oma';
-import fetchOrthologsFromOrthodb from './orthodb';
+import {
+  fetchOrthologsFromOrthodb, fetchOrthologsFromOrthodbSparql
+} from './orthodb';
 
 async function fetchOrthologs(genes, sourceOrg, targetOrgs, api='oma') {
   if (api === 'oma') {
     return await fetchOrthologsFromOma(genes, sourceOrg, targetOrgs);
   } else if (api === 'orthodb') {
-    return await fetchOrthologsFromOrthodb(genes, sourceOrg, targetOrgs);
+    return await fetchOrthologsFromOrthodbSparql(genes, sourceOrg, targetOrgs);
   }
 }