--- a/src/homology.js
+++ b/src/homology.js
@@ -1,13 +1,13 @@
 import fetchOrthologsFromOma from './oma';
 import {
-  fetchOrthologsFromOrthodb, fetchOrthologsFromOrthodbSparql
+  fetchOrthologsFromOrthodbSparql
 } from './orthodb';
 
-async function fetchOrthologs(genes, sourceOrg, targetOrgs, api='oma') {
-  if (api === 'oma') {
-    return await fetchOrthologsFromOma(genes, sourceOrg, targetOrgs);
-  } else if (api === 'orthodb') {
+async function fetchOrthologs(genes, sourceOrg, targetOrgs, api='orthodb') {
+  if (api === 'orthodb') {
     return await fetchOrthologsFromOrthodbSparql(genes, sourceOrg, targetOrgs);
+  } else if (api === 'oma') {
+    return await fetchOrthologsFromOma(genes, sourceOrg, targetOrgs);
   }
 }