This switches Homology to use the Orthodb SPARQL API instead of its REST API.  It also switches to using MyGene.info to fetch gene locations, and only calls NCBI when that fails.

The result is a substantially more robust and faster Homology API.

| Query | Old time (s) | New time (s) |
|-------|-------------|--------------|
| MAPT | 3.5 | 1.7 |
| APP | Error (3.4) *** | 6.7 |
| GTF2B, SIX5, NR2C2, E2F4, NFYA, BRD2,<br/>RBPJ, GTF2F1, NFYC, KDM5A, STAT2, SP2 | 6.8 | 6.1 |
| RAD51,RAD51C,XRCC3,DMC1,RAD51D,<br/>XRCC2,RAD51B,MAPK11,BRCA2,RAD52,ABL1,RAD54L | N/A (9.3) *  |  6.0 |
| MTHFR,UROD,BHMT2,BHMT,MTR | N/A (4.9) ** | 3.7 |

\* Old API displayed `TypeError: Cannot read property 'chrloc' of undefined`
\** Old API displayed `Error: Gene "RAD51B" not found in source organism "Homo sapiens"`
\*** Old API displayed `Error: Gene "BHMT" not found in source organism "Homo sapiens"`